+++ pclsync/putil.h
@@ -6,12 +6,15 @@ extern "C" {
 #endif
 
 #include <stddef.h>
+#include <time.h>
 
 #define NTO_STR(s) TO_STR(s)
 #define TO_STR(s) #s
 #define VAR_ARRAY(name, type, size) type name[size]
+#define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
 
 void putil_wipe(void *mem, size_t sz);
+void time_format(time_t tm, unsigned long ns, char *result);
 
 #ifdef __cplusplus
 }
